requirements.txt
Django==4.2.19
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
coreapi==2.3.3
markdown==3.5.2
python-dateutil==2.8.2
pytz==2024.1
dj-database-url==2.1.0
gunicorn==21.2.0
django-leaflet==0.28.0
psycopg2-binary==2.9.9
